Added GlycoBlocks toggle Refactored folder structure Reduced cube size for mtz file upload
