+++ webapp/src/database/DatabaseComponents/BFactorVsRSCC.tsx
