+++ webapp/src/database/DatabaseComponents/SNFGSection.tsx
