+++ webapp/src/database/DatabaseComponents/SugarList.tsx
@@ -1,34 +1,58 @@
 import React, { useMemo, useEffect, useState } from 'react';
 import { useTable } from 'react-table';
-import { DatabaseColumns } from '../../data/Constants';
+import { SugarListColumns } from '../../data/Constants.tsx';
 import styled from 'styled-components';
 
-function customSort(a, b): number {
-    a = a.chain;
-    b = b.chain;
-    if (a < b) return -1;
-    if (a > b) return 1;
+function customSort(
+    a: Record<string, string>,
+    b: Record<string, string>
+): number {
+    const splitA = a['Sugar ID'].split('-');
+    const splitB = b['Sugar ID'].split('-');
+
+    if (splitA[1] < splitB[1]) {
+        return -1;
+    }
+    if (splitB[1] < splitA[1]) {
+        return 1;
+    }
+
+    if (Number(splitA[2]) < Number(splitB[2])) {
+        return -1;
+    }
+
+    if (Number(splitB[2]) < Number(splitA[2])) {
+        return 1;
+    }
+
     return 0;
 }
 
 function parseResults(data) {
     const glycans = data.data.glycans;
 
-    const tableData: any[] = [];
+    const tableData = [];
 
     for (const key in glycans) {
         const glycanType = glycans[key];
         for (let i = 0; i < glycanType.length; i++) {
-            const chain = glycanType[i].RootSugarChainID;
+            const sugars: Array<Record<string, any>> = glycanType[i].Sugars;
+            for (let j = 0; j < sugars.length; j++) {
+                const keys = Object.keys(sugars[j]);
+                const entry = {};
 
-            const SNFG = glycanType[i].SNFG;
-            const WURCS = glycanType[i].WURCS;
+                for (let k = 0; k < keys.length; k++) {
+                    const typeKey = typeof sugars[j][keys[k]];
 
-            tableData.push({
-                chain,
-                SNFG,
-                WURCS,
-            });
+                    if (typeKey === 'number') {
+                        entry[keys[k]] = sugars[j][keys[k]].toFixed(2);
+                    } else {
+                        entry[keys[k]] = sugars[j][keys[k]];
+                    }
+                }
+                entry.type = key;
+                tableData.push(entry);
+            }
         }
     }
 
@@ -95,9 +119,9 @@ const Styles = styled.div`
     }
 `;
 
-export default function SNFGList(props) {
-    const [data, setData] = useState<any[]>([]);
-    const columns = useMemo(() => DatabaseColumns, []);
+export default function SugarList(props) {
+    const [data, setData] = useState([]);
+    const columns = useMemo(() => SugarListColumns, []);
     const {
         _getTableProps,
         getTableBodyProps,
@@ -112,9 +136,9 @@ export default function SNFGList(props) {
     }, []);
 
     return (
-        <div className="flex flex-col mx-auto px-16">
+        <div className="flex flex-col mx-auto p-16">
             <span className="text-xl mb-2">
-                N- and O-glycan structure 2D descriptions
+                Detailed monosaccharide validation data
             </span>
 
             <Styles>
