+++ webapp/src/main/GlycanDetail/GlycanDetailMoorhenView.tsx
@@ -9,6 +9,7 @@ export function GlycanDetailMoorhenView(props: {
     moorhenProps: any;
     moorhenDimensions: () => [number, number];
     mapContour: number;
+    toggleGlycoblocks: () => any;
 }) {
     const size = Math.min(800, 0.9 * window.innerWidth);
     const [dimensions, setDimensions] = useState<Record<string, number>>(
@@ -113,6 +114,35 @@ export function GlycanDetailMoorhenView(props: {
                         ></div>
                         <span className="ms-3 text-sm font-medium text-gray-900 dark:text-gray-300">
                             Toggle Symmetry
+                        </span>
+                    </label>
+                </div>
+
+                <div className="mt-4">
+                    <label className="relative inline-flex items-center cursor-pointer">
+                        <input
+                            type="checkbox"
+                            value=""
+                            className="sr-only peer"
+                            onMouseDown={(e) => {
+                                e.stopPropagation();
+                            }}
+                            onTouchStart={(e) => {
+                                e.stopPropagation();
+                            }}
+                        />
+                        <div
+                            onMouseDown={(e) => {
+                                e.stopPropagation();
+                            }}
+                            onTouchStart={(e) => {
+                                e.stopPropagation();
+                            }}
+                            onClick={props.toggleGlycoblocks}
+                            className="w-11 h-6 bg-tertiary peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-blue-300 dark:peer-focus:ring-blue-800 rounded-full peer dark:bg-gray-700 peer-checked:after:translate-x-full rtl:peer-checked:after:-translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:start-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all dark:border-gray-600 peer-checked:bg-blue-600"
+                        ></div>
+                        <span className="ms-3 text-sm font-medium text-gray-900 dark:text-gray-300">
+                            Toggle Glycoblocks
                         </span>
                     </label>
                 </div>
