+++ webapp/src/main/GlycanDetail/GlycanDetailTable.jsx
